series index reset + minor fixes
