+++ tests/operator/transform/test_op_update.py
@@ -12,6 +12,11 @@ from openclean.function.value.mapping import Lookup
 from openclean.operator.stream.collector import Collector
 from openclean.operator.transform.update import swap, update, Update
 
+from openclean.function.matching.fuzzy import FuzzySimilarity
+from openclean.function.matching.base import DefaultStringMatcher
+from openclean.function.matching.mapping import best_matches
+from openclean.function.value.domain import BestMatch
+
 
 BOROUGHS = [
     'BRONX',
@@ -107,3 +112,39 @@ def test_unary_update_consumer():
     assert rows[0] == (3, [1, 1, 3])
     assert rows[1] == (2, [4, 4, 6])
     assert rows[2] == (1, [7, 7, 9])
+
+
+def test_update_bestmatch_column(nyc311):
+    """Test updating values in a single column with their best match mappings
+    of a data frame.
+    """
+    domain_pronounciation = [
+        'bronks',
+        'brooklen',
+        'manhatn',
+        'kweenz',
+        'staten islen'
+    ]
+
+    vocabulary = DefaultStringMatcher(
+        vocabulary=domain_pronounciation,
+        similarity=FuzzySimilarity(),
+        no_match_threshold=0.
+    )
+
+    # replace with the best match in the given vocabulary
+    df = update(nyc311, 'borough', BestMatch(matcher=vocabulary))
+    values = df['borough'].unique()
+    assert len(values) == 5
+    assert sorted(values) == sorted(domain_pronounciation)
+
+    # Manually create a mapping to let user manipulate it before updating the
+    # data frame.
+    map = best_matches(values=nyc311['borough'].unique(), matcher=vocabulary)
+    # override best match from the vocabulary
+    map.update({'STATEN ISLAND': 'statn ayeln'})
+
+    df = update(nyc311, 'borough', map.to_lookup())
+    values = df['borough'].unique()
+    assert len(values) == 5
+    assert sorted(values) == sorted(map.to_lookup().values())
